dz1

+++ src/dzMain.java
@@ -4,7 +4,7 @@ public class dzMain {
     public static void main(String[] args){
         Scanner in = new Scanner(System.in);
         System.out.print("Введите первое число : ");
-        int a = in.nextInt();
+        double a = in.nextInt();
 
         Scanner sc = new Scanner(System.in);
         System.out.print("Введите знак : ");
@@ -12,7 +12,7 @@ public class dzMain {
 
         Scanner lo = new Scanner(System.in);
         System.out.print("Введите второе число : ");
-        int b = lo.nextInt();
+        double b = lo.nextInt();
 
         sc.close();
         System.out.print("Ответ : ");
